Adding others Business rules

+++ Teams/APP.Layer/CQRS/Handlers/DeleteTeamHandler.cs
@@ -1,11 +1,14 @@
 using MediatR;
 using Teams.API.Layer.Middlewares;
 using Teams.APP.Layer.CQRS.Commands;
+using Teams.APP.Layer.Helpers;
+using Teams.CORE.Layer.BusinessExceptions;
 using Teams.CORE.Layer.Interfaces;
 
 namespace Teams.APP.Layer.CQRS.Handlers;
 
-public class DeleteTeamHandler(ITeamRepository teamRepository) : IRequestHandler<DeleteTeamCommand>
+public class DeleteTeamHandler(ITeamRepository teamRepository, ILogger<DeleteTeamHandler> log)
+    : IRequestHandler<DeleteTeamCommand>
 {
     public async Task Handle(DeleteTeamCommand command, CancellationToken cancellationToken)
     {
@@ -17,6 +20,21 @@ public class DeleteTeamHandler(ITeamRepository teamRepository) : IRequestHandler
                 "Not Found",
                 "Team ID not found"
             );
+        var result = team.IsProjectHasAnyDependencies();
+        if (result)
+        {
+            LogHelper.BusinessRuleFailure(
+                log,
+                "Delete Team",
+                $"🚫 The team {team.Name} has been associated to 1 or more projects.",
+                null
+            );
+
+            throw new DomainException(
+                $"The team {team.Name} has been associated to 1 or more projects."
+            );
+        }
         await teamRepository.DeleteTeamAsync(command.Id);
+        LogHelper.Info($"✅ Team with Name {team.Name} has been deleted successfully.", log);
     }
 }
